MyPy update. Avoid re-doing work; avoid decoding twice. Make compression methods static and associated with model. When the decoding process changes, so should the extract and compress methods.
